fix: ActiveWalletProviderでSmartWallet初期化を単一化
useActiveWallet を Context 化し、複数コンポーネントで重複していた
SmartAccount 初期化を 1 回に集約。isSmartWallet が embedded 検出後に
正しく true に遷移するよう安定化。
